W-11558572 - Moved OASComponentConfiguration to OASConfiguration
