+++ amf-api-contract/shared/src/main/scala/amf/apicontract/client/platform/AMFConfiguration.scala
@@ -177,7 +177,17 @@ object RAMLConfiguration {
 object OASConfiguration {
   def OAS20(): AMFConfiguration              = InternalOASConfiguration.OAS20()
   def OAS30(): AMFConfiguration              = InternalOASConfiguration.OAS30()
+  def OAS30Component(): AMFConfiguration     = InternalOASConfiguration.OAS30Component()
+
+  /**
+   * common configuration to parse OAS20 and OAS30 APIs
+   */
   def OAS(): AMFConfiguration                = InternalOASConfiguration.OAS()
+
+  /**
+   * common configuration to parse OAS30Component fragments
+   */
+  def OASComponent(): AMFConfiguration       = InternalOASConfiguration.OASComponent()
   def fromSpec(spec: Spec): AMFConfiguration = InternalOASConfiguration.fromSpec(spec)
 }
